+++ lib/list/body/header.h
@@ -17,7 +17,7 @@ bool clu_list_body_insert(list_body_p *lb_root, handler_p h);
 bool clu_list_body_remove(list_body_p *lb_root, handler_p h);
 
 uint64_t clu_list_body_count(list_body_p lb);
-handler_p clu_list_body_get_handler(list_body_p lb, int y);
+handler_p clu_list_body_get_handler(list_body_p lb, uint64_t y);
 bool clu_list_body_contains(list_body_p lb, handler_p h);
 
 #endif
