+++ lib/mem/header.h
@@ -4,9 +4,10 @@
 #ifdef DEBUG
 
 #include <stdbool.h>
+#include <stdbool.h>
 
-extern int list_head_alive;
-extern int list_body_alive;
+extern uint64_t list_head_alive;
+extern uint64_t list_body_alive;
 
 void clu_mem_internal_display();
 bool clu_mem_internal_empty();
